Added the generation of Players, including the corresponding Gameboard
Every Cell on a Gameboard has coordinates, and can be manipulated.

The Avaiable states of a Cell are as follows:
-Empty (No boolean is true)
-Hit
-Occupied (by a ship part)
-Hit and Occupied (both booleans are true)

An example of how to manipulate the singular Cells is inside the generatePlayers Function inside the Main Controller

To Generate the Players, one has to push the "Start Long Game" Button, then the Text Version of the Gameboard will be printed to the console.